processor/requirements.txt
pandas>=2.0.0
tables>=3.8.0
openpyxl>=3.1.0
telethon>=1.30.0
python-socks[asyncio]>=2.0.0
tqdm
pyyaml
